Added method to allow user to search for vehicles based on price range. Added method to print the List of vehicles

+++ CarDealershipADV/src/main/java/Data/VehicleDao.java
@@ -8,7 +8,7 @@ public interface VehicleDao {
 
 	List<Vehicle> getAll();
 
-	List<Vehicle> searchByPrice(int minPrice, int maxPrice);
+	List<Vehicle> searchByPrice(double minPrice, double maxPrice);
 
 	List<Vehicle> searchByMake(String make);
